+++ Application/Services/OrderService.cs
@@ -3,6 +3,7 @@ using Common.Exceptions;
 using Domain.Interfaces;
 using Domain.Models.Entities;
 using Domain.Models.RequestModels;
+using Domain.Models.ResponseModels;
 using FluentValidation;
 using Microsoft.Extensions.Logging;
 
@@ -21,16 +22,22 @@ namespace Application.Services
             _orderValidator = orderValidator;
         }
 
-        public async Task<IEnumerable<Order>> GetAllOrders()
+        public async Task<IEnumerable<OrderResponseModel>> GetAllOrders()
         {
             _logger.LogInformation("Retrieving all orders");
             var orders = await _orderRepository.GetAllOrdersAsync();
 
             _logger.LogInformation("Retrieved {OrderCount} orders", orders.Count());
-            return orders;
+            return orders.Select(order => new OrderResponseModel
+            {
+                OrderId = order.Id,
+                OrderDate = order.OrderDate,
+                TotalValue = order.TotalValue,
+                ClientId = order.ClientId
+            });
         }
 
-        public async Task<Order> GetOrderById(int id)
+        public async Task<OrderResponseModel> GetOrderById(int id)
         {
             _logger.LogInformation("Starting order search with ID {Id}", id);
             var order = await _orderRepository.GetOrderByIdAsync(id);
@@ -42,10 +49,16 @@ namespace Application.Services
             }
 
             _logger.LogInformation("Order found by ID: {OrderId}", order.Id);
-            return order;
+            return new OrderResponseModel()
+            {
+                OrderId = order.Id,
+                OrderDate = order.OrderDate,
+                TotalValue = order.TotalValue,
+                ClientId = order.ClientId
+            }; 
         }
 
-        public async Task<Order> CreateOrder(OrderRequestModel orderRequestModel)
+        public async Task<OrderResponseModel> CreateOrder(OrderRequestModel orderRequestModel)
         {
             _logger.LogInformation("Starting order creation with request data: {OrderRequest}", orderRequestModel);
             var order = new Order
@@ -66,13 +79,29 @@ namespace Application.Services
             await _orderRepository.CreateAsync(order);
 
             _logger.LogInformation("Order created with ID: {OrderId}", order.Id);
-            return order;
+            return new OrderResponseModel()
+            {
+                OrderId = order.Id,
+                OrderDate = order.OrderDate,
+                TotalValue = order.TotalValue,
+                ClientId = order.ClientId
+            };
         }
 
         public async Task UpdateOrder(int id, OrderRequestModel orderRequestModel)
         {
             _logger.LogInformation("Starting order update with request data: {OrderRequest}", orderRequestModel);
-            var order = await GetOrderById(id);
+
+            _logger.LogInformation("Starting order search with ID {Id}", id);
+            var order = await _orderRepository.GetOrderByIdAsync(id);
+
+            if (order == null)
+            {
+                _logger.LogError("Order not found by ID: {Id}", id);
+                throw new NotFoundException($"Order not found by ID: {id}");
+            }
+
+            _logger.LogInformation("Order found by ID: {OrderId}", order.Id);
 
             order.OrderDate = orderRequestModel.OrderDate;
             order.TotalValue = orderRequestModel.TotalValue;
@@ -94,7 +123,16 @@ namespace Application.Services
         {
             _logger.LogInformation("Deleting order with ID: {Id}", id);
 
-            var order = await GetOrderById(id);
+            _logger.LogInformation("Starting order search with ID {Id}", id);
+            var order = await _orderRepository.GetOrderByIdAsync(id);
+
+            if (order == null)
+            {
+                _logger.LogError("Order not found by ID: {Id}", id);
+                throw new NotFoundException($"Order not found by ID: {id}");
+            }
+
+            _logger.LogInformation("Order found by ID: {OrderId}", order.Id);
             await _orderRepository.DeleteAsync(order);
         }
     }
